Refactor dependency resolution out of deployment handler
